Improve build_graph: faster and shorter code
Improve build_graph() a lot by completely rewriting it and using the
points coordinates as nodes. By that the code does not need to compare
the start and end points of the different routes to find duplicates.

This results in the dependency to numpy to fall away for
find_shortest_path.py and util.py and the complexity to change from n^2
to n.